Move logic from NewEntry to DataEntryTable, and other progress.
